adapt post type providers to php-di
