+++ wp-content/plugins/core/definitions/post-types.php
@@ -1,0 +1,4 @@
+<?php
+declare( strict_types=1 );
+
+return [];
