+++ wp-content/plugins/core/src/Post_Types/Sample/Config.php
@@ -6,6 +6,11 @@ namespace Tribe\Project\Post_Types\Sample;
 use Tribe\Libs\Post_Type\Post_Type_Config;
 
 class Config extends Post_Type_Config {
+	protected $post_type = Sample::NAME;
+
+	public function __construct() {
+	}
+
 	public function get_args() {
 		return [
 			'hierarchical'     => false,
