--- wp-content/plugins/core/src/Service_Providers/Post_Types/Sample_Service_Provider.php
@@ -1,11 +1,0 @@
-<?php
-
-
-namespace Tribe\Project\Service_Providers\Post_Types;
-
-use Tribe\Project\Post_Types\Sample;
-
-class Sample_Service_Provider extends Post_Type_Service_Provider {
-	protected $post_type_class = Sample\Sample::class;
-	protected $config_class = Sample\Config::class;
-}
